+++ pygluu/containerlib/constants.py
@@ -1,3 +1,10 @@
+"""
+pygluu.containerlib.constants
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+This module contains values that are not supposedly modified.
+"""
+
 PERSISTENCE_TYPES = (
     "ldap",
     "couchbase",